Introduce new exception QemuImgNotFound
At the moment the exception QemuNotFound is used when the binary
qemu-img is not available or if QEMU is not installed. There are
distributions packaging qemu-img in a separate package. So it is
possible to have qemu-img installed without installed QEMU. To
reflect this behavior a new exception QemuImgNotFound exception
should be used when qemu-img was not found.